first pass refactor account blockchain scanner service away
