+++ libs/blockchain/src/blockchain-scanner.service.spec.ts
@@ -1,26 +1,31 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { Test } from '@nestjs/testing';
+import { Test, TestingModule } from '@nestjs/testing';
 import { BlockHash, BlockNumber, SignedBlock } from '@polkadot/types/interfaces';
-import { InjectRedis } from '@songkeys/nestjs-redis';
+import { getRedisToken, InjectRedis } from '@songkeys/nestjs-redis';
 import { Redis } from 'ioredis';
 import { BlockchainScannerService } from './blockchain-scanner.service';
 import { FrameSystemEventRecord } from '@polkadot/types/lookup';
 import { mockApiPromise } from '#testlib/polkadot-api.mock.spec';
-import { BlockchainRpcQueryService } from '#blockchain/blockchain-rpc-query.service';
 import blockchainConfig, { IBlockchainNonProviderConfig } from '#blockchain/blockchain.config';
 import { GenerateMockConfigProvider } from '#testlib/utils.config-tests';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-import { AnyNumber } from '@polkadot/types/types';
-import { LoggerModule, PinoLogger } from 'nestjs-pino';
+import { InjectPinoLogger, LoggerModule } from 'nestjs-pino';
 import { getPinoHttpOptions } from '#logger-lib';
 import { mockRedisProvider } from '#testlib';
+import { BlockchainService, NONCE_SERVICE_REDIS_NAMESPACE } from '#blockchain/blockchain.service';
+import { AnyNumber } from '@polkadot/types/types';
 
 jest.mock('@polkadot/api', () => {
   const originalModule = jest.requireActual<typeof import('@polkadot/api')>('@polkadot/api');
   return {
     __esModules: true,
-    WsProvider: jest.fn().mockImplementation(() => originalModule.WsProvider),
+    WsProvider: jest.fn().mockImplementation(() => ({
+      disconnect: jest.fn(),
+      ...originalModule.WsProvider,
+    })),
     ApiPromise: jest.fn().mockImplementation(() => ({
+      consts: {},
+      disconnect: jest.fn(),
       ...originalModule.ApiPromise,
       ...mockApiPromise,
     })),
@@ -30,8 +35,15 @@ jest.mock('@polkadot/api', () => {
 const mockBlockchainConfigProvider = GenerateMockConfigProvider<IBlockchainNonProviderConfig>(blockchainConfig.KEY, {
   frequencyTimeoutSecs: 10,
   frequencyApiWsUrl: new URL('ws://localhost:9944'),
-  isDeployedReadOnly: false,
+  isDeployedReadOnly: false
 });
+
+const mockedRedisProvider = mockRedisProvider();
+
+const mockNonceRedisProvider = {
+  provide: getRedisToken(NONCE_SERVICE_REDIS_NAMESPACE),
+  useValue: mockedRedisProvider.useValue,
+};
 
 const mockBlockHash = {
   toString: jest.fn(() => '0x1234'),
@@ -58,8 +70,8 @@ const mockEmptyBlockHash = {
 class ScannerService extends BlockchainScannerService {
   constructor(
     @InjectRedis() redis: Redis,
-    blockchainService: BlockchainRpcQueryService,
-    protected readonly logger: PinoLogger,
+    blockchainService: BlockchainService,
+    @InjectPinoLogger('ScannerService') protected readonly logger,
   ) {
     super(redis, blockchainService, logger);
   }
@@ -71,11 +83,13 @@ class ScannerService extends BlockchainScannerService {
 
 describe('BlockchainScannerService', () => {
   let service: ScannerService;
-  let blockchainService: BlockchainRpcQueryService;
+  let blockchainService: BlockchainService;
+  let moduleRef: TestingModule;
 
   beforeAll(async () => {
-    const moduleRef = await Test.createTestingModule({
+    moduleRef = await Test.createTestingModule({
       imports: [
+        LoggerModule.forRoot(getPinoHttpOptions()),
         EventEmitterModule.forRoot({
           // Use this instance throughout the application
           global: true,
@@ -94,17 +108,20 @@ describe('BlockchainScannerService', () => {
           // disable throwing uncaughtException if an error event is emitted and it has no listeners
           ignoreErrors: false,
         }),
-        LoggerModule.forRoot(getPinoHttpOptions()),
       ],
-      providers: [mockRedisProvider(), mockBlockchainConfigProvider, Logger, BlockchainRpcQueryService, ScannerService],
+      providers: [
+        mockedRedisProvider,
+        mockNonceRedisProvider,
+        mockBlockchainConfigProvider,
+        Logger,
+        BlockchainService,
+        ScannerService,
+      ],
     }).compile();
-
     moduleRef.enableShutdownHooks();
-
     service = moduleRef.get<ScannerService>(ScannerService);
-    blockchainService = moduleRef.get<BlockchainRpcQueryService>(BlockchainRpcQueryService);
+    blockchainService = moduleRef.get<BlockchainService>(BlockchainService);
     const mockApi: any = await blockchainService.getApi();
-
     jest.spyOn(blockchainService, 'getBlock').mockResolvedValue(mockSignedBlock as unknown as SignedBlock);
     jest
       .spyOn(blockchainService, 'getBlockHash')
@@ -113,8 +130,11 @@ describe('BlockchainScannerService', () => {
       );
     jest.spyOn(blockchainService, 'getLatestBlockNumber');
     jest.spyOn(blockchainService, 'getEvents').mockResolvedValue([]);
-
     mockApi.emit('connected'); // keeps the test suite from hanging when finished
+  });
+
+  afterAll(async () => {
+    moduleRef.close();
   });
 
   describe('scan', () => {
